requirements.txt
pandas
scikit-learn
scipy
numpy
scikit-surprise
